+++ src/actions/warnings.js
@@ -1,14 +1,23 @@
 import { toast } from "react-toastify";
 import { ToastSuccess } from "../components/ui/ToastSuccess";
+import { fetchWithToken } from "../helpers/fetch";
 import { types } from "../types/types";
 
 export const startLoadWarnings = () => {
 
     return async (dispatch) => {
 
-        // fecth warnings
+        try {
+            // fecth warnings
+            const resp = await fetchWithToken("warnings/");
+            const { warnings } = await resp.json();
 
-        dispatch(loadWarnings(mockWarnings));
+            if (warnings) {
+                dispatch(loadWarnings(warnings));
+            }
+        } catch (error) {
+            console.log(error);
+        }
 
     }
 }
@@ -22,16 +31,20 @@ export const startDeleteWarning = (id) => {
 
     return async (dispatch) => {
 
-        // TODO: delete id from DB
+        try {
+            // Delete warning from DB
+            const resp = await fetchWithToken(`warnings/${id}`, {}, 'DELETE');
+            const { destroyed } = await resp.json();
 
+            if (destroyed) {
+                dispatch(deleteWarning(id));
+                toast.success(<ToastSuccess text="Aviso eliminado con éxito" />);
+            }
+        } catch (error) {
+            console.log(error);
+        }
 
-
-        dispatch(deleteWarning(id));
-        setTimeout(() => {
-            toast.success(<ToastSuccess text="Aviso eliminado con éxito" />);
-        }, 600);
     }
-
 }
 
 const deleteWarning = (id) => ({
@@ -39,22 +52,26 @@ const deleteWarning = (id) => ({
     payload: id
 });
 
-export const startAddWarning = (warning) => {
+export const startAddWarning = ({ description }) => {
 
     return async (dispatch) => {
 
-        // TODO: save warning to DB
-        const newWarning = {
-            id: new Date().getTime(),
-            description: warning.description
+        try {
+            // Save warning to DB
+            const resp = await fetchWithToken("warnings/", { description }, 'POST');
+            const { warning } = await resp.json();
+
+            if (warning) {
+                setTimeout(() => {
+                    dispatch(addWarning({ id: warning.id, description: warning.description }));
+                    toast.success(<ToastSuccess text="Aviso agregado con éxito" />);
+                }, 600);
+
+            }
+        } catch (error) {
+            console.log(error);
         }
-        dispatch(addWarning(newWarning));
-
-        setTimeout(() => {
-            toast.success(<ToastSuccess text="Aviso agregado con éxito" />);
-        }, 600);
     }
-
 }
 
 const addWarning = (warning) => ({
@@ -65,118 +82,3 @@ const addWarning = (warning) => ({
 export const clearWarnings = () => ({
     type: types.warningClear
 });
-
-const mockWarnings = [
-
-    {
-        id: '12313123',
-        description: 'F1 H3 parará 17/03/2020 17:50 5 horas'
-    },
-    {
-        id: '12315123a',
-        description: 'F2 P3 paro 19/04/2020 18:50-20:45'
-    },
-    {
-        id: '12321323c',
-        description: 'Pedir electrodos'
-    }, {
-        id: '12313123v',
-        description: 'F1 H3 parará 17/03/2020 17:50 5 horas'
-    },
-    {
-        id: '12315123b',
-        description: 'F2 P3 paro 19/04/2020 18:50-20:45'
-    },
-    {
-        id: '12321323n',
-        description: 'Pedir electrodos'
-    }, {
-        id: '12313123m',
-        description: 'F1 H3 parará 17/03/2020 17:50 5 horas'
-    },
-    {
-        id: '123151233v',
-        description: 'F2 P3 paro 19/04/2020 18:50-20:45'
-    },
-    {
-        id: '12321323v',
-        description: 'Pedir electrodos'
-    }, {
-        id: '12313123a',
-        description: 'F1 H3 parará 17/03/2020 17:50 5 horas'
-    },
-    {
-        id: '12315123v',
-        description: 'F2 P3 paro 19/04/2020 18:50-20:45'
-    },
-    {
-        id: '12321323sa',
-        description: 'Pedir electrodos'
-    }, {
-        id: '12313123vc',
-        description: 'F1 H3 parará 17/03/2020 17:50 5 horas'
-    },
-    {
-        id: '12315123as',
-        description: 'F2 P3 paro 19/04/2020 18:50-20:45'
-    },
-    {
-        id: '12321323csdfgvsd',
-        description: 'Pedir electrodos'
-    }, {
-        id: '12313123s',
-        description: 'F1 H3 parará 17/03/2020 17:50 5 horas'
-    },
-    {
-        id: '12315123vsd',
-        description: 'F2 P3 paro 19/04/2020 18:50-20:45'
-    },
-    {
-        id: '12321323abf',
-        description: 'Pedir electrodos'
-    }, {
-        id: '12313123vdas',
-        description: 'F1 H3 parará 17/03/2020 17:50 5 horas'
-    },
-    {
-        id: '12315123s',
-        description: 'F2 P3 paro 19/04/2020 18:50-20:45'
-    },
-    {
-        id: '12321323vvsd',
-        description: 'Pedir electrodos'
-    }, {
-        id: '12313123avdasv',
-        description: 'F1 H3 parará 17/03/2020 17:50 5 horas'
-    },
-    {
-        id: '12315123vnrty',
-        description: 'F2 P3 paro 19/04/2020 18:50-20:45'
-    },
-    {
-        id: '12321323sdqwqwe',
-        description: 'Pedir electrodos'
-    }, {
-        id: '12313123gfwergb',
-        description: 'F1 H3 parará 17/03/2020 17:50 5 horas'
-    },
-    {
-        id: '12315123scdv',
-        description: 'F2 P3 paro 19/04/2020 18:50-20:45'
-    },
-    {
-        id: '12321323vsdv',
-        description: 'Pedir electrodos'
-    }, {
-        id: '12313123brgfb',
-        description: 'F1 H3 parará 17/03/2020 17:50 5 horas'
-    },
-    {
-        id: '12315123qwdq',
-        description: 'F2 P3 paro 19/04/2020 18:50-20:45'
-    },
-    {
-        id: '12321323svddv',
-        description: 'Pedir electrodos'
-    },
-]
